iceberg_rest

Module ax_lib

Modules

body

HTTP body utilities.

error_handling

Error handling model and utilities

extract

Types and traits for extracting data from requests.

handler

Async functions that can be used to handle requests.

http

A general purpose library of common HTTP types

middleware

Utilities for writing middleware

response

Types and traits for generating responses.

routing

Routing between [Service]s and handlers.

serve

Serve services.

Structs

Error

Errors that can happen when using axum.

Extension

Extractor and response for extensions.

Form

URL encoded extractor and response.

Json

JSON Extractor / Response.

Router

The router type for composing handlers and services.

Traits

RequestExt

Extension trait that adds additional methods to Request.

RequestPartsExt

Extension trait that adds additional methods to Parts.

ServiceExt

Extension trait that adds additional methods to any [Service].

Functions

serve

Serve the service with the supplied listener.

Type Aliases

BoxError

Alias for a type-erased error type.

Attribute Macros

async_trait