iceberg_rest::ax_lib::extract

Struct Query

pub struct Query<T>(pub T);

Extractor that deserializes query strings into some type.

T is expected to implement serde::Deserialize.

Example

use axum::{
    extract::Query,
    routing::get,
    Router,
};
use serde::Deserialize;

#[derive(Deserialize)]
struct Pagination {
    page: usize,
    per_page: usize,
}

// This will parse query strings like `?page=2&per_page=30` into `Pagination`
// structs.
async fn list_things(pagination: Query<Pagination>) {
    let pagination: Pagination = pagination.0;

    // ...
}

let app = Router::new().route("/list_things", get(list_things));

If the query string cannot be parsed it will reject the request with a 400 Bad Request response.

For handling values being empty vs missing see the query-params-with-empty-strings example.

For handling multiple values for the same query parameter, in a ?foo=1&foo=2&foo=3 fashion, use axum_extra::extract::Query instead.

Tuple Fields
0: T
Implementations

impl<T> Query<T>
where T: DeserializeOwned,

pub fn try_from_uri(value: &Uri) -> Result<Query<T>, QueryRejection>

Attempts to construct a Query from a reference to a Uri.

Example

use axum::extract::Query;
use http::Uri;
use serde::Deserialize;

#[derive(Deserialize)]
struct ExampleParams {
    foo: String,
    bar: u32,
}

let uri: Uri = "http://example.com/path?foo=hello&bar=42".parse().unwrap();
let result: Query<ExampleParams> = Query::try_from_uri(&uri).unwrap();
assert_eq!(result.foo, String::from("hello"));
assert_eq!(result.bar, 42);