iceberg_rest::ax_lib::routing::method_routing

Struct MethodRouter

pub struct MethodRouter<S = (), E = Infallible> {
    get: MethodEndpoint<S, E>,
    head: MethodEndpoint<S, E>,
    delete: MethodEndpoint<S, E>,
    options: MethodEndpoint<S, E>,
    patch: MethodEndpoint<S, E>,
    post: MethodEndpoint<S, E>,
    put: MethodEndpoint<S, E>,
    trace: MethodEndpoint<S, E>,
    connect: MethodEndpoint<S, E>,
    fallback: Fallback<S, E>,
    allow_header: AllowHeader,
}

A [Service] that accepts requests based on a MethodFilter and allows chaining additional handlers and services.

When does MethodRouter implement [Service]?

Whether or not MethodRouter implements [Service] depends on the state type it requires.

use tower::Service;
use axum::{routing::get, extract::{State, Request}, body::Body};

// this `MethodRouter` doesn't require any state, i.e. the state is `()`,
let method_router = get(|| async {});
// and thus it implements `Service`
assert_service(method_router);

// this requires a `String` and doesn't implement `Service`
let method_router = get(|_: State<String>| async {});
// until you provide the `String` with `.with_state(...)`
let method_router_with_state = method_router.with_state(String::new());
// and then it implements `Service`
assert_service(method_router_with_state);

// helper to check that a value implements `Service`
fn assert_service<S>(service: S)
where
    S: Service<Request>,
{}

Fields
get: MethodEndpoint<S, E>head: MethodEndpoint<S, E>delete: MethodEndpoint<S, E>options: MethodEndpoint<S, E>patch: MethodEndpoint<S, E>post: MethodEndpoint<S, E>put: MethodEndpoint<S, E>trace: MethodEndpoint<S, E>connect: MethodEndpoint<S, E>fallback: Fallback<S, E>allow_header: AllowHeader
Implementations

impl<S> MethodRouter<S>
where S: Clone,

pub fn on<H, T>(self, filter: MethodFilter, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will accept requests matching the given MethodFilter.

Example

use axum::{
    routing::get,
    Router,
    routing::MethodFilter
};

async fn handler() {}

async fn other_handler() {}

// Requests to `GET /` will go to `handler` and `DELETE /` will go to
// `other_handler`
let app = Router::new().route("/", get(handler).on(MethodFilter::DELETE, other_handler));

pub fn connect<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept CONNECT requests.

See MethodFilter::CONNECT for when you’d want to use this, and MethodRouter::get for an example.

pub fn delete<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept DELETE requests.

See MethodRouter::get for an example.

pub fn get<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept GET requests.

Example

use axum::{routing::post, Router};

async fn handler() {}

async fn other_handler() {}

// Requests to `POST /` will go to `handler` and `GET /` will go to
// `other_handler`.
let app = Router::new().route("/", post(handler).get(other_handler));

Note that get routes will also be called for HEAD requests but will have the response body removed. Make sure to add explicit HEAD routes afterwards.

pub fn head<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept HEAD requests.

See MethodRouter::get for an example.

pub fn options<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept OPTIONS requests.

See MethodRouter::get for an example.

pub fn patch<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept PATCH requests.

See MethodRouter::get for an example.

pub fn post<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept POST requests.

See MethodRouter::get for an example.

pub fn put<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept PUT requests.

See MethodRouter::get for an example.

pub fn trace<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Chain an additional handler that will only accept TRACE requests.

See MethodRouter::get for an example.

pub fn fallback<H, T>(self, handler: H) -> MethodRouter<S>
where H: Handler<T, S>, T: 'static, S: Send + Sync + 'static,

Add a fallback Handler to the router.

impl MethodRouter

pub fn into_make_service(self) -> IntoMakeService<MethodRouter>

Convert the router into a MakeService.

This allows you to serve a single MethodRouter if you don’t need any routing based on the path:

use axum::{
    handler::Handler,
    http::{Uri, Method},
    response::IntoResponse,
    routing::get,
};
use std::net::SocketAddr;

async fn handler(method: Method, uri: Uri, body: String) -> String {
    format!("received `{method} {uri}` with body `{body:?}`")
}

let router = get(handler).post(handler);

let listener = tokio::net::TcpListener::bind("0.0.0.0:3000").await.unwrap();
axum::serve(listener, router.into_make_service()).await.unwrap();

pub fn into_make_service_with_connect_info<C>( self, ) -> IntoMakeServiceWithConnectInfo<MethodRouter, C>

Convert the router into a MakeService which stores information about the incoming connection.

See Router::into_make_service_with_connect_info for more details.

use axum::{
    handler::Handler,
    response::IntoResponse,
    extract::ConnectInfo,
    routing::get,
};
use std::net::SocketAddr;

async fn handler(ConnectInfo(addr): ConnectInfo<SocketAddr>) -> String {
    format!("Hello {addr}")
}

let router = get(handler).post(handler);

let listener = tokio::net::TcpListener::bind("0.0.0.0:3000").await.unwrap();
axum::serve(listener, router.into_make_service()).await.unwrap();

impl<S, E> MethodRouter<S, E>
where S: Clone,

pub fn new() -> MethodRouter<S, E>

Create a default MethodRouter that will respond with 405 Method Not Allowed to all requests.

pub fn with_state<S2>(self, state: S) -> MethodRouter<S2, E>

Provide the state for the router.

pub fn on_service<T>(self, filter: MethodFilter, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will accept requests matching the given MethodFilter.

Example

use axum::{
    extract::Request,
    Router,
    routing::{MethodFilter, on_service},
    body::Body,
};
use http::Response;
use std::convert::Infallible;

let service = tower::service_fn(|request: Request| async {
    Ok::<_, Infallible>(Response::new(Body::empty()))
});

// Requests to `DELETE /` will go to `service`
let app = Router::new().route("/", on_service(MethodFilter::DELETE, service));

pub fn connect_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept CONNECT requests.

See MethodFilter::CONNECT for when you’d want to use this, and MethodRouter::get_service for an example.

pub fn delete_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept DELETE requests.

See MethodRouter::get_service for an example.

pub fn get_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept GET requests.

Example

use axum::{
    extract::Request,
    Router,
    routing::post_service,
    body::Body,
};
use http::Response;
use std::convert::Infallible;

let service = tower::service_fn(|request: Request| async {
    Ok::<_, Infallible>(Response::new(Body::empty()))
});

let other_service = tower::service_fn(|request: Request| async {
    Ok::<_, Infallible>(Response::new(Body::empty()))
});

// Requests to `POST /` will go to `service` and `GET /` will go to
// `other_service`.
let app = Router::new().route("/", post_service(service).get_service(other_service));

Note that get routes will also be called for HEAD requests but will have the response body removed. Make sure to add explicit HEAD routes afterwards.

pub fn head_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept HEAD requests.

See MethodRouter::get_service for an example.

pub fn options_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept OPTIONS requests.

See MethodRouter::get_service for an example.

pub fn patch_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept PATCH requests.

See MethodRouter::get_service for an example.

pub fn post_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept POST requests.

See MethodRouter::get_service for an example.

pub fn put_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept PUT requests.

See MethodRouter::get_service for an example.

pub fn trace_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Chain an additional service that will only accept TRACE requests.

See MethodRouter::get_service for an example.

pub fn fallback_service<T>(self, svc: T) -> MethodRouter<S, E>
where T: Service<Request<Body>, Error = E> + Clone + Send + 'static, <T as Service<Request<Body>>>::Response: IntoResponse + 'static, <T as Service<Request<Body>>>::Future: Send + 'static,

Add a fallback service to the router.

This service will be called if no routes matches the incoming request.

use axum::{
    Router,
    routing::get,
    handler::Handler,
    response::IntoResponse,
    http::{StatusCode, Method, Uri},
};

let handler = get(|| async {}).fallback(fallback);

let app = Router::new().route("/", handler);

async fn fallback(method: Method, uri: Uri) -> (StatusCode, String) {
    (StatusCode::NOT_FOUND, format!("`{method}` not allowed for {uri}"))
}

When used with MethodRouter::merge

Two routers that both have a fallback cannot be merged. Doing so results in a panic:

use axum::{
    routing::{get, post},
    handler::Handler,
    response::IntoResponse,
    http::{StatusCode, Uri},
};

let one = get(|| async {}).fallback(fallback_one);

let two = post(|| async {}).fallback(fallback_two);

let method_route = one.merge(two);

async fn fallback_one() -> impl IntoResponse { /* ... */ }
async fn fallback_two() -> impl IntoResponse { /* ... */ }

Setting the Allow header

By default MethodRouter will set the Allow header when returning 405 Method Not Allowed. This is also done when the fallback is used unless the response generated by the fallback already sets the Allow header.

This means if you use fallback to accept additional methods, you should make sure you set the Allow header correctly.

pub fn layer<L, NewError>(self, layer: L) -> MethodRouter<S, NewError>
where L: Layer<Route<E>> + Clone + Send + 'static, <L as Layer<Route<E>>>::Service: Service<Request<Body>> + Clone + Send + 'static, <<L as Layer<Route<E>>>::Service as Service<Request<Body>>>::Response: IntoResponse + 'static, <<L as Layer<Route<E>>>::Service as Service<Request<Body>>>::Error: Into<NewError> + 'static, <<L as Layer<Route<E>>>::Service as Service<Request<Body>>>::Future: Send + 'static, E: 'static, S: 'static, NewError: 'static,

Apply a [tower::Layer] to all routes in the router.

This can be used to add additional processing to a request for a group of routes.

Note that the middleware is only applied to existing routes. So you have to first add your routes (and / or fallback) and then call layer afterwards. Additional routes added after layer is called will not have the middleware added.

Works similarly to Router::layer. See that method for more details.

Example

use axum::{routing::get, Router};
use tower::limit::ConcurrencyLimitLayer;

async fn handler() {}

let app = Router::new().route(
    "/",
    // All requests to `GET /` will be sent through `ConcurrencyLimitLayer`
    get(handler).layer(ConcurrencyLimitLayer::new(64)),
);

pub fn route_layer<L>(self, layer: L) -> MethodRouter<S, E>
where L: Layer<Route<E>> + Clone + Send + 'static, <L as Layer<Route<E>>>::Service: Service<Request<Body>, Error = E> + Clone + Send + 'static, <<L as Layer<Route<E>>>::Service as Service<Request<Body>>>::Response: IntoResponse + 'static, <<L as Layer<Route<E>>>::Service as Service<Request<Body>>>::Future: Send + 'static, E: 'static, S: 'static,

Apply a [tower::Layer] to the router that will only run if the request matches a route.

Note that the middleware is only applied to existing routes. So you have to first add your routes (and / or fallback) and then call route_layer afterwards. Additional routes added after route_layer is called will not have the middleware added.

This works similarly to MethodRouter::layer except the middleware will only run if the request matches a route. This is useful for middleware that return early (such as authorization) which might otherwise convert a 405 Method Not Allowed into a 401 Unauthorized.

Example

use axum::{
    routing::get,
    Router,
};
use tower_http::validate_request::ValidateRequestHeaderLayer;

let app = Router::new().route(
    "/foo",
    get(|| async {})
        .route_layer(ValidateRequestHeaderLayer::bearer("password"))
);

// `GET /foo` with a valid token will receive `200 OK`
// `GET /foo` with a invalid token will receive `401 Unauthorized`
// `POST /FOO` with a invalid token will receive `405 Method Not Allowed`

pub fn merge(self, other: MethodRouter<S, E>) -> MethodRouter<S, E>

Merge two routers into one.

This is useful for breaking routers into smaller pieces and combining them into one.

use axum::{
    routing::{get, post},
    Router,
};

let get = get(|| async {});
let post = post(|| async {});

let merged = get.merge(post);

let app = Router::new().route("/", merged);

// Our app now accepts
// - GET /
// - POST /

pub fn handle_error<F, T>(self, f: F) -> MethodRouter<S>
where F: Clone + Send + Sync + 'static, HandleError<Route<E>, F, T>: Service<Request<Body>, Error = Infallible>, <HandleError<Route<E>, F, T> as Service<Request<Body>>>::Future: Send, <HandleError<Route<E>, F, T> as Service<Request<Body>>>::Response: IntoResponse + Send, T: 'static, E: 'static, S: 'static,

Apply a HandleErrorLayer.

This is a convenience method for doing self.layer(HandleErrorLayer::new(f)).